Add guest can view an item
Add route and view for visitor viewing an item. Added image and description to item table as they would likely not be driven out by tests in the immediate future.

Encountered a small bug in test for viewing all items related to the item's name which didn't exist yet in the factory file.